Move Colors and Style to core folder
