+++ components/core/Style.tsx
